Rename NI to new directory naming convention
