+++ feature/networkinstance/networkinstance_test.go
@@ -14,7 +14,7 @@
  limitations under the License.
 */
 
-package nibase
+package networkinstance
 
 import (
 	"errors"
